move cached rest mapper out of the sdk
