+++ main.go
@@ -19,8 +19,8 @@ import (
 	"os"
 
 	"github.com/banzaicloud/logging-operator/controllers"
+	"github.com/banzaicloud/logging-operator/pkg/k8sutil"
 	loggingv1alpha2 "github.com/banzaicloud/logging-operator/pkg/sdk/api/v1beta1"
-	"github.com/banzaicloud/logging-operator/pkg/sdk/util"
 	prometheusOperator "github.com/coreos/prometheus-operator/pkg/apis/monitoring/v1"
 	apiextensions "k8s.io/apiextensions-apiserver/pkg/apis/apiextensions/v1beta1"
 	"k8s.io/apimachinery/pkg/runtime"
@@ -61,7 +61,7 @@ func main() {
 		Scheme:             scheme,
 		MetricsBindAddress: metricsAddr,
 		LeaderElection:     enableLeaderElection,
-		MapperProvider:     util.NewCached,
+		MapperProvider:     k8sutil.NewCached,
 	})
 
 	if err != nil {
