+++ pkg/k8sutil/cache.go
@@ -4,17 +4,17 @@
 // you may not use this file except in compliance with the License.
 // You may obtain a copy of the License at
 //
-//    http://www.apache.org/licenses/LICENSE-2.0
+//     http://www.apache.org/licenses/LICENSE-2.0
 //
 // Unless required by applicable law or agreed to in writing, software
 // distributed under the License is distributed on an "AS IS" BASIS,
 // WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
 // See the License for the specific language governing permissions and
 // limitations under the License.
-//
+
 // This code is based on https://raw.githubusercontent.com/kubernetes-sigs/cluster-api/ee96c31eea0f9b1b2f4cffae5c7da2c274722e76/util/restmapper/cached.go
 
-package util
+package k8sutil
 
 import (
 	"sync"
