+++ scripts/check-header.sh
@@ -17,7 +17,7 @@ read -r -d '' EXPECTED <<EOF
 EOF
 
 STATUS=0
-FILES=$(find . -name "*.go" -not -path ./pkg/sdk/util/cache.go)
+FILES=$(find . -name "*.go" -not -path ./pkg/k8sutil/cache.go)
 
 for FILE in $FILES; do
     # Skip check for autogenerated files
